+++ SofascoreAcademy/SofascoreAcademy/Extensions/UICollectionViewExtensions.swift
@@ -1,0 +1,9 @@
+import Foundation
+import UIKit
+
+extension UICollectionView {
+    
+    func scrollToCenterForItem(at indexPath: IndexPath, animated: Bool) {
+        scrollToItem(at: indexPath, at: .centeredHorizontally, animated: animated)
+    }
+}
