+++ SofascoreAcademy/SofascoreAcademy/Network/ImageService.swift
@@ -4,7 +4,7 @@ import UIKit
 class ImageService {
     
     private let imageCache = ImageCache.shared
-
+    
     func getLeagueLogo(tournamentId: Int) async -> UIImage {
         
         let cacheKey: String = "league_\(tournamentId)"
