--- SofascoreAcademy/SofascoreAcademy/Protocols/TeamLogoLoadProtocol.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol TeamLogoLoadProtocol: AnyObject {
-    
-    func fetchTeamsLogosFromApi(homeTeamId: Int, awayTeamId: Int, indexPath: IndexPath)
-}
