+++ SofascoreAcademy/SofascoreAcademy/Public/Styling.swift
@@ -7,7 +7,7 @@ public enum Fonts {
     static let RobotoBold16 = UIFont(name: "Roboto-Bold", size: 16)
     static let RobotoBold20 = UIFont(name: "Roboto-Bold", size: 20)
     static let RobotoBold32 = UIFont(name: "Roboto-Bold", size: 32)
-
+    
     static let RobotoRegular14 = UIFont(name: "Roboto-Regular", size: 14)
     static let RobotoRegular12 = UIFont(name: "Roboto-Regular", size: 12)
     static let RobotoCondensedRegularMicro = UIFont(name: "RobotoCondensed-Regular", size: 12)
@@ -24,96 +24,96 @@ public enum Colors {
 
 //ovo sam kopirao sa zeplina, ali mi neki fontovi ne rade pa sam ostavio i ovo i moj Enum gore :/
 extension UIFont {
-  class var headline1Desktop: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 32.0)!
-  }
-  class var headline1: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 20.0)!
-  }
-  class var headline2: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 16.0)!
-  }
-  class var action: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 16.0)!
-  }
-  class var headline3: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 14.0)!
-  }
-  class var tabular: UIFont {
-    return UIFont(name: "Roboto-CondensedRegular", size: 14.0)!
-  }
-  class var bodyParagraph: UIFont {
-    return UIFont(name: "Roboto-Regular", size: 14.0)!
-  }
-  class var body: UIFont {
-    return UIFont(name: "Roboto-Regular", size: 14.0)!
-  }
-  class var assistive: UIFont {
-    return UIFont(name: "Roboto-Bold", size: 12.0)!
-  }
-  class var micro: UIFont {
-    return UIFont(name: "Roboto-CondensedRegular", size: 12.0)!
-  }
+    class var headline1Desktop: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 32.0)!
+    }
+    class var headline1: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 20.0)!
+    }
+    class var headline2: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 16.0)!
+    }
+    class var action: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 16.0)!
+    }
+    class var headline3: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 14.0)!
+    }
+    class var tabular: UIFont {
+        return UIFont(name: "Roboto-CondensedRegular", size: 14.0)!
+    }
+    class var bodyParagraph: UIFont {
+        return UIFont(name: "Roboto-Regular", size: 14.0)!
+    }
+    class var body: UIFont {
+        return UIFont(name: "Roboto-Regular", size: 14.0)!
+    }
+    class var assistive: UIFont {
+        return UIFont(name: "Roboto-Bold", size: 12.0)!
+    }
+    class var micro: UIFont {
+        return UIFont(name: "Roboto-CondensedRegular", size: 12.0)!
+    }
 }
 
 extension UIColor {
-  @nonobjc class var colorPrimaryVariant: UIColor {
-    return UIColor(red: 19.0 / 255.0, green: 39.0 / 255.0, blue: 186.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var colorSecondaryVariant: UIColor {
-    return UIColor(red: 221.0 / 255.0, green: 219.0 / 255.0, blue: 201.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var colorPrimaryDefault: UIColor {
-    return UIColor(red: 55.0 / 255.0, green: 77.0 / 255.0, blue: 245.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var colorSecondaryDefault: UIColor {
-    return UIColor(red: 240.0 / 255.0, green: 238.0 / 255.0, blue: 223.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var colorPrimaryHighlight: UIColor {
-    return UIColor(red: 225.0 / 255.0, green: 237.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var colorSecondaryHighlight: UIColor {
-    return UIColor(red: 247.0 / 255.0, green: 246.0 / 255.0, blue: 239.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var onColorOnColorPrimary: UIColor {
-    return UIColor(white: 1.0, alpha: 1.0)
-  }
-  @nonobjc class var surfaceSurface1: UIColor {
-    return UIColor(white: 1.0, alpha: 1.0)
-  }
-  @nonobjc class var surfaceSurface0: UIColor {
-    return UIColor(red: 239.0 / 255.0, green: 243.0 / 255.0, blue: 248.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var surfaceSurface2: UIColor {
-    return UIColor(red: 192.0 / 255.0, green: 207.0 / 255.0, blue: 228.0 / 255.0, alpha: 0.2)
-  }
-  @nonobjc class var onSurfaceOnSurfaceLv1: UIColor {
-    return UIColor(white: 18.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var onColorOnColorSecondary: UIColor {
-    return UIColor(white: 1.0, alpha: 0.6)
-  }
-  @nonobjc class var onSurfaceOnSurfaceLv2: UIColor {
-    return UIColor(white: 18.0 / 255.0, alpha: 0.4)
-  }
-  @nonobjc class var onSurfaceOnSurfaceLv3: UIColor {
-    return UIColor(white: 18.0 / 255.0, alpha: 0.2)
-  }
-  @nonobjc class var onSurfaceOnSurfaceLv4: UIColor {
-    return UIColor(white: 18.0 / 255.0, alpha: 0.1)
-  }
-  @nonobjc class var statusError: UIColor {
-    return UIColor(red: 234.0 / 255.0, green: 69.0 / 255.0, blue: 69.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var statusAlert: UIColor {
-    return UIColor(red: 216.0 / 255.0, green: 177.0 / 255.0, blue: 39.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var statusSuccess: UIColor {
-    return UIColor(red: 29.0 / 255.0, green: 168.0 / 255.0, blue: 109.0 / 255.0, alpha: 1.0)
-  }
-  @nonobjc class var specificLive: UIColor {
-    return UIColor(red: 233.0 / 255.0, green: 48.0 / 255.0, blue: 48.0 / 255.0, alpha: 1.0)
-  }
+    @nonobjc class var colorPrimaryVariant: UIColor {
+        return UIColor(red: 19.0 / 255.0, green: 39.0 / 255.0, blue: 186.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var colorSecondaryVariant: UIColor {
+        return UIColor(red: 221.0 / 255.0, green: 219.0 / 255.0, blue: 201.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var colorPrimaryDefault: UIColor {
+        return UIColor(red: 55.0 / 255.0, green: 77.0 / 255.0, blue: 245.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var colorSecondaryDefault: UIColor {
+        return UIColor(red: 240.0 / 255.0, green: 238.0 / 255.0, blue: 223.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var colorPrimaryHighlight: UIColor {
+        return UIColor(red: 225.0 / 255.0, green: 237.0 / 255.0, blue: 254.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var colorSecondaryHighlight: UIColor {
+        return UIColor(red: 247.0 / 255.0, green: 246.0 / 255.0, blue: 239.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var onColorOnColorPrimary: UIColor {
+        return UIColor(white: 1.0, alpha: 1.0)
+    }
+    @nonobjc class var surfaceSurface1: UIColor {
+        return UIColor(white: 1.0, alpha: 1.0)
+    }
+    @nonobjc class var surfaceSurface0: UIColor {
+        return UIColor(red: 239.0 / 255.0, green: 243.0 / 255.0, blue: 248.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var surfaceSurface2: UIColor {
+        return UIColor(red: 192.0 / 255.0, green: 207.0 / 255.0, blue: 228.0 / 255.0, alpha: 0.2)
+    }
+    @nonobjc class var onSurfaceOnSurfaceLv1: UIColor {
+        return UIColor(white: 18.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var onColorOnColorSecondary: UIColor {
+        return UIColor(white: 1.0, alpha: 0.6)
+    }
+    @nonobjc class var onSurfaceOnSurfaceLv2: UIColor {
+        return UIColor(white: 18.0 / 255.0, alpha: 0.4)
+    }
+    @nonobjc class var onSurfaceOnSurfaceLv3: UIColor {
+        return UIColor(white: 18.0 / 255.0, alpha: 0.2)
+    }
+    @nonobjc class var onSurfaceOnSurfaceLv4: UIColor {
+        return UIColor(white: 18.0 / 255.0, alpha: 0.1)
+    }
+    @nonobjc class var statusError: UIColor {
+        return UIColor(red: 234.0 / 255.0, green: 69.0 / 255.0, blue: 69.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var statusAlert: UIColor {
+        return UIColor(red: 216.0 / 255.0, green: 177.0 / 255.0, blue: 39.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var statusSuccess: UIColor {
+        return UIColor(red: 29.0 / 255.0, green: 168.0 / 255.0, blue: 109.0 / 255.0, alpha: 1.0)
+    }
+    @nonobjc class var specificLive: UIColor {
+        return UIColor(red: 233.0 / 255.0, green: 48.0 / 255.0, blue: 48.0 / 255.0, alpha: 1.0)
+    }
 }
